requirements.txt
boto3 
PyYAML